+++ doc/changelog.md
@@ -8,6 +8,7 @@
 
 ## Deprecation
 - The isWindows() function is deprecated in favour of the IS_WINDOWS constant (which is faster). 
+- BuildContext.mergeOptions() is deprecated in favour of BaseTarget.options, or (for situations where there is no target such as PathSets) BuildContext.getGlobalOption().
 
 ## Fixes
 - Fixed a couple of bugs in incremental C++ compilation - one that could cause unnecessary incremental compilation of targets that depend on generated C/C++ source or include files, and another in which the build would fail rather than re-running makedepends if some of the cached dependencies no longer exist.